+++ internal/config/types.go
@@ -1,6 +1,9 @@
 package config
 
-const ErrMetaTypeEmpty string = "meta type empty"
+const (
+	ErrMetaTypeEmpty         string = "meta type empty"
+	ErrConfigurationNotFound string = "no configuration found"
+)
 
 const (
 	DefaultConfigFile string = "bludgeon.json"