+++ internal/kafka/client.go
@@ -202,21 +202,20 @@ func (k *kafka) Configure(items ...interface{}) error {
 	k.Lock()
 	defer k.Unlock()
 
-	var envs map[string]string
 	var c *Configuration
 
 	for _, item := range items {
 		switch v := item.(type) {
 		case config.Envs:
-			envs = v
+			c = new(Configuration)
+			c.Default()
+			c.FromEnv(v)
 		case *Configuration:
 			c = v
 		}
 	}
 	if c == nil {
-		c = new(Configuration)
-		c.Default()
-		c.FromEnv(envs)
+		return errors.New(config.ErrConfigurationNotFound)
 	}
 	if err := c.Validate(); err != nil {
 		return err