+++ internal/meta/mysql/types.go
@@ -1,0 +1,8 @@
+package mysql
+
+import "database/sql"
+
+const (
+	DatabaseIsolation = sql.LevelSerializable
+	LogAlias          = "[mysql_client]"
+)