+++ internal/websocket/client/config.go
@@ -1,7 +1,6 @@
 package client
 
 import (
-	"strconv"
 	"time"
 
 	"github.com/pkg/errors"
@@ -10,11 +9,6 @@ import (
 const (
 	ReadTimeout  string = "read timeout is less than 0"
 	WriteTimeout string = "read timeout is less than 0"
-)
-
-const (
-	EnvNameReadTimeout  string = "BLUDGEON_WEBSOCKET_READ_TIMEOUT"
-	EnvNameWriteTimeout string = "BLUDGEON_WEBSOCKET_WRITE_TIMEOUT"
 )
 
 const (
@@ -30,36 +24,4 @@ var (
 type Configuration struct {
 	WriteTimeout time.Duration `json:"write_timeout"`
 	ReadTimeout  time.Duration `json:"read_timeout"`
-}
-
-func (c *Configuration) Default() {
-	c.ReadTimeout = DefaultReadTimeout
-	c.WriteTimeout = DefaultWriteTimeout
-}
-
-func (c *Configuration) FromEnv(envs map[string]string) {
-	if readTimeoutString, ok := envs[EnvNameReadTimeout]; ok {
-		if readtimeoutInt, err := strconv.Atoi(readTimeoutString); err == nil {
-			if readTimeout := time.Duration(readtimeoutInt) * time.Second; readTimeout > 0 {
-				c.ReadTimeout = readTimeout
-			}
-		}
-	}
-	if writeTimeoutString, ok := envs[EnvNameWriteTimeout]; ok {
-		if writeTimeoutInt, err := strconv.Atoi(writeTimeoutString); err == nil {
-			if writeTimeout := time.Duration(writeTimeoutInt) * time.Second; writeTimeout > 0 {
-				c.WriteTimeout = writeTimeout
-			}
-		}
-	}
-}
-
-func (c *Configuration) Validate() (err error) {
-	if c.ReadTimeout < 0 {
-		return ErrReadTimeout
-	}
-	if c.WriteTimeout < 0 {
-		return ErrWriteTimeout
-	}
-	return
 }