+++ internal/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2023-05-05
+
+- added cache package
+- added queue package
+- streamlined dependency management such that nested pointers were neither initialized or parameterized
+
 ## [1.4.3] - 2023-02-22
 
 - upgraded to golang.org/x/net v0.7.0